+++ spec/support/active_record/models.rb
@@ -5,11 +5,16 @@ module Models
   module ActiveRecord
     class User < ::ActiveRecord::Base
       audited except: :password
-      attribute :non_column_attr if Rails.version >= "5.1"
+      attribute :non_column_attr if Rails.gem_version >= Gem::Version.new("5.1")
       attr_protected :logins if respond_to?(:attr_protected)
-      enum status: {active: 0, reliable: 1, banned: 2}
 
-      if Rails.version >= "7.1"
+      if Rails.gem_version >= Gem::Version.new("7.2")
+        enum :status, {active: 0, reliable: 1, banned: 2}
+      else
+        enum status: {active: 0, reliable: 1, banned: 2}
+      end
+
+      if Rails.gem_version >= Gem::Version.new("7.1")
         serialize :phone_numbers, type: Array
       else
         serialize :phone_numbers, Array
@@ -27,7 +32,7 @@ module Models
 
     class UserOnlyPassword < ::ActiveRecord::Base
       self.table_name = :users
-      attribute :non_column_attr if Rails.version >= "5.1"
+      attribute :non_column_attr if Rails.gem_version >= Gem::Version.new("5.1")
       audited only: :password
     end
 
@@ -52,6 +57,12 @@ module Models
         audited
         encrypts :password
       end
+    end
+
+    class UserWithReadOnlyAttrs < ::ActiveRecord::Base
+      self.table_name = :users
+      audited
+      attr_readonly :status
     end
 
     class CommentRequiredUser < ::ActiveRecord::Base
@@ -130,6 +141,12 @@ module Models
       has_associated_audits
       has_many :companies, class_name: "OwnedCompany", dependent: :destroy
       accepts_nested_attributes_for :companies
+
+      if Rails.gem_version >= Gem::Version.new("7.2")
+        enum :status, {active: 0, reliable: 1, banned: 2}
+      else
+        enum status: {active: 0, reliable: 1, banned: 2}
+      end
     end
 
     class OwnedCompany < ::ActiveRecord::Base