+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Audited ChangeLog
 
+### 5.7.0 (2024-08-13)
+
+- Support for Rails 7.2 and Ruby 3.3, and testing cleanups - @mattbrictson
+  [#723](https://github.com/collectiveidea/audited/pull/723)
+- Allow max_audits to be a proc or symbol - @gmhawash
+  [#718](https://github.com/collectiveidea/audited/pull/718)
+- Support Rails 8 - @fernandomenolli
+  [#717](https://github.com/collectiveidea/audited/pull/717)
+
+
+### 5.6.0 (2024-04-05)
+
+- Removed support for Rails 5.0 and 5.1.
+- Replace RequestStore with ActiveSupport::CurrentAttributes - @punkisdead
+  [#702](https://github.com/collectiveidea/audited/pull/702)
+
+### 5.5.0 (2024-04-02)
+
+- Bad release. Same code as 5.4.1. Use 5.6.0 for updated features.
+
+### 5.4.3 (2024-01-11)
+
+- Ignore readonly columns in audit - @sriddbs
+  [#692](https://github.com/collectiveidea/audited/pull/692)
+- Robustify Rails version checks - @blaet
+  [#689](https://github.com/collectiveidea/audited/pull/689)
+-  Ignore callbacks if not specifed on the model
+  [#679](https://github.com/collectiveidea/audited/pull/679)
+
 ## 5.4.2 (2023-11-30)
 
 - Revert replacing RequetStore with ActiveSupport::CurrentAttributes until it is fully tested.
@@ -19,7 +48,7 @@
 ## 5.3.3 (2023-03-24)
 
 - Use RequestStore instead of Thread.current for thread-safe requests - @tiagocassio
-  [#669](https://github.com/collectiveidea/audited/pull/669)
+  [#669](https://github.com/c ollectiveidea/audited/pull/669)
 - Clean up Touch audits - @mcyoung, @akostadinov
   [#668](https://github.com/collectiveidea/audited/pull/668)